Backport of #4516 (#4721)
