+++ distributor/go.mod
@@ -6,10 +6,9 @@ require (
 	github.com/cloudevents/sdk-go/protocol/nats/v2 v2.4.1
 	github.com/cloudevents/sdk-go/v2 v2.4.1
 	github.com/kelseyhightower/envconfig v1.4.0
-	github.com/keptn/go-utils v0.8.5
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/nats-io/nats-server/v2 v2.2.6
-	github.com/nats-io/nats.go v1.11.0
-	github.com/sirupsen/logrus v1.4.2
+	github.com/keptn/go-utils v0.8.6-0.20210710122139-586c8979384d
+	github.com/nats-io/nats-server/v2 v2.3.1
+	github.com/nats-io/nats.go v1.11.1-0.20210623165838-4b75fc59ae30
+	github.com/sirupsen/logrus v1.8.1
 	github.com/stretchr/testify v1.7.0
 )
