+++ distributor/pkg/lib/controlplane/controlplane.go
@@ -1,4 +1,4 @@
-package lib
+package controlplane
 
 import (
 	"fmt"
