+++ distributor/pkg/lib/controlplane/uniformlog.go
@@ -1,4 +1,4 @@
-package lib
+package controlplane
 
 import (
 	"context"
