+++ distributor/pkg/lib/controlplane/uniformlog_test.go
@@ -1,4 +1,4 @@
-package lib
+package controlplane
 
 import (
 	"context"
