+++ distributor/pkg/lib/events/nats_connection_handler_test.go
@@ -1,4 +1,4 @@
-package lib
+package events
 
 import (
 	"fmt"
