+++ distributor/pkg/lib/events/utils.go
@@ -1,9 +1,75 @@
-package lib
+package events
 
 import (
+	"context"
+	"encoding/json"
+	cloudevents "github.com/cloudevents/sdk-go/v2"
 	keptnmodels "github.com/keptn/go-utils/pkg/api/models"
+	"github.com/keptn/go-utils/pkg/common/sliceutils"
+	"github.com/keptn/go-utils/pkg/lib/v0_2_0"
+	"github.com/keptn/keptn/distributor/pkg/config"
+	logger "github.com/sirupsen/logrus"
+	"strings"
 	"sync"
 )
+
+type EventMatcher struct {
+	Project string
+	Stage   string
+	Service string
+}
+
+func NewEventMatcherFromEnv(config config.EnvConfig) *EventMatcher {
+	return &EventMatcher{
+		Project: config.ProjectFilter,
+		Stage:   config.StageFilter,
+		Service: config.ServiceFilter,
+	}
+}
+
+func (ef EventMatcher) Matches(e cloudevents.Event) bool {
+	keptnBase := &v0_2_0.EventData{}
+	if err := e.DataAs(keptnBase); err != nil {
+		return true
+	}
+	if ef.Project != "" && !sliceutils.ContainsStr(strings.Split(ef.Project, ","), keptnBase.Project) ||
+		ef.Stage != "" && !sliceutils.ContainsStr(strings.Split(ef.Stage, ","), keptnBase.Stage) ||
+		ef.Service != "" && !sliceutils.ContainsStr(strings.Split(ef.Service, ","), keptnBase.Service) {
+		return false
+	}
+	return true
+}
+
+type ExecutionContext struct {
+	context.Context
+	Wg *sync.WaitGroup
+}
+
+func NewExecutionContext(ctx context.Context, waitGroupCount int) *ExecutionContext {
+	wg := new(sync.WaitGroup)
+	wg.Add(waitGroupCount)
+	return &ExecutionContext{ctx, wg}
+}
+
+type ceVersion struct {
+	SpecVersion string `json:"specversion"`
+}
+
+func DecodeCloudEvent(data []byte) (*cloudevents.Event, error) {
+	cv := &ceVersion{}
+	if err := json.Unmarshal(data, cv); err != nil {
+		return nil, err
+	}
+
+	event := cloudevents.NewEvent(cv.SpecVersion)
+
+	if err := json.Unmarshal(data, &event); err != nil {
+		logger.Errorf("Could not unmarshal CloudEvent: %v", err)
+		return nil, err
+	}
+
+	return &event, nil
+}
 
 type CloudEventsCache struct {
 	sync.RWMutex
